move components into UserProfile
