--- frontend/src/app/components/ProfileHeader.module.css
@@ -1,67 +1,0 @@
-.profileHeader {
-  display: flex;
-  width: 1024px;
-  justify-content: space-between;
-}
-
-.profileContent {
-  display: flex;
-  gap: 24px;
-}
-
-.userInfo {
-  display: flex;
-  flex-direction: column;
-  gap: var(--Spacing-s, 8px);
-  justify-content: center;
-}
-
-.userInfoHeader {
-  display: flex;
-  align-items: center;
-  gap: var(--Spacing-sm, 16px);
-}
-
-.userFullName {
-  color: var(--Text-3, #3a3c3d);
-
-  /* title medium */
-  font-family: "Albert Sans";
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-
-.userMetadata {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: var(--Spacing-s, 8px);
-  flex-wrap: wrap;
-}
-
-.userMetadata .metadataSubsection {
-  display: flex;
-}
-
-.userMetadata .metadataSubsection * {
-  width: fit-content;
-  color: var(--Text-1, #60696f);
-  /* body large */
-  font-family: "Albert Sans";
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-
-.userMetadata .smallMetadata {
-  font-size: 14px;
-}
-
-.profileContentControls {
-  display: flex;
-  padding: var(--Spacing-s, 8px) 0px;
-  gap: var(--Spacing-s, 8px);
-}
